q2/requirements.txt
Django==2.0.4
django-rest-swagger==2.1.2
djangorestframework==3.8.2
simplejson==3.13.2
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.4.0